feat: workContainerUtil 파일 분리
